refactor file conversion and progress streaming into class
