+++ src/controllers/conversion/index.model.ts
@@ -1,4 +1,4 @@
-interface File extends Blob {
+export interface File extends Blob {
     readonly lastModified: number;
     readonly name: string;
     readonly mimetype: string;
