+++ src/controllers/conversion/index.ts
@@ -1,2 +1,4 @@
 export * from "./index.controller";
 export * from "./index.events";
+export * from "./index.middleware";
+export * from "./index.model";
